NASS-853: edit mode initial values properly handled
